support/visual-builder/backend/requirements.txt
fastapi
uvicorn
requests
psutil
nomos[cli,serve,traces,openai,mistralai,gemini,ollama,huggingface]
python-multipart
python-dotenv
